pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"
requires_python = ">=3.10"

[project]
name = "hardhat-event-streams"
version = "0.0.1"
authors = [{name = "Matt Krueger", email = "mkrueger@rstms.net"}]
readme = {file = "README.md", content-type = "text/markdown"}
license = {file = "LICENSE"}
keywords = ["hardhat evm contract event streams"]
classifiers = [
    "Intended Audience :: Developers",
    "Natural Language :: English", 
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
  ]
dynamic = ["description"]

dependencies = [
    "fastapi",
    "uvicorn",
    "sqlmodel",
    "eth-utils"
  ]

[tool.flit.module]
name = "hardhat_event_streams"

[project.optional-dependencies]
dev = [
    "black",
    "bump2version",
    "coverage",
    "flake8",
    "flake8-length",
    "flit",
    "isort",
    "pdbpp",
    "pytest",
    "pytest-asyncio",
    "pytest-datadir",
    "httpx"
  ]
docs = [
    "m2r2",
    "sphinx",
    "sphinx-click",
    "sphinx-rtd-theme"
  ]

[project.urls]
  Home = "https://github.com/rstms/hardhat-event-streams"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[project.scripts]
hes = "hardhat_event_streams.cli:cli"
